Speed up Playwright e2e CI
